make sure menu values are not limited to 8bits

+++ app/inc/naveg.h
@@ -143,7 +143,7 @@ void naveg_print_pb_name(uint8_t display);
 
 void naveg_lock_pots(uint8_t lock);
 
-void naveg_menu_item_changed_cb(uint8_t item_ID, uint8_t value);
+void naveg_menu_item_changed_cb(uint8_t item_ID, uint16_t value);
 
 uint8_t naveg_banks_mode_pb(void);
